requirements.txt
langchain==0.1.16
langchain_community==0.0.32
langchain_openai==0.1.3
PyPDF2==3.0.1
python-dotenv==1.0.1
streamlit==1.32.2
streamlit_extras==0.4.2
faiss-cpu==1.8.0
